feat: shared backgrounds
