--- .storybook/preview.js
@@ -1,15 +1,0 @@
-export const parameters = {
-  backgrounds: {
-    default: "light",
-    values: [
-      {
-        name: "light",
-        value: "#ffffff",
-      },
-      {
-        name: "dark",
-        value: "#000000",
-      },
-    ],
-  },
-};
