+++ packages/icons/.storybook/preview.ts
@@ -1,0 +1,18 @@
+// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+import { backgroundDark, backgroundLight } from "../../material/src";
+
+export const parameters = {
+  backgrounds: {
+    default: "light",
+    values: [
+      {
+        name: "light",
+        value: backgroundLight,
+      },
+      {
+        name: "dark",
+        value: backgroundDark,
+      },
+    ],
+  },
+};
